requirements.txt
streamlit
yfinance
pandas
numpy
plotly
scikit-learn
transformers
torch
arch
newsapi-python